refactor(investment): split InvestmentCharts.tsx into section components (CHAOS-1227)
